Implement saveArticle function for handling article saving logic and localStorage management
- Added saveArticle function to manage saving articles with title, body, and images.
- Implemented localStorage and sessionStorage comparison for title and body.
- Added logic to handle image storage and comparison between session and local storage.
- Created RELOAD ARTICLE and SAVE PROCESS documentation for clarity on new functionalities.
- Updated DashboardEditor to utilize saveArticle for auto-saving drafts.
- Refactored handleClick to integrate new saveArticle logic.
- Adjusted saveImageTemporally to include imageId for better identification.
- Cleaned up console logs for production readiness.
